Move ast definitions around, rename stuff slightly
