+++ src/main/kotlin/com/github/derg/transpiler/core/Print.kt
@@ -2,12 +2,11 @@ package com.github.derg.transpiler.core
 
 import com.github.derg.transpiler.source.Name
 import com.github.derg.transpiler.source.ast.*
-import com.github.derg.transpiler.source.ast.Function
 
 /**
  * Pretty-prints the entire abstract syntax tree defined from [this] node.
  */
-fun Node.print() = println(prettify())
+fun Segment.print() = println(prettify())
 
 /**
  * Converts [this] node into a string representation of the tree. Child nodes are included in the string on new lines
@@ -24,11 +23,13 @@ private fun Node.prettify(): String = when (this)
     is Assignment.AssignModulo   -> stringifyAssignment("%=", name, expression)
     is Assignment.AssignMultiply -> stringifyAssignment("*=", name, expression)
     is Assignment.AssignSubtract -> stringifyAssignment("-=", name, expression)
-    is Control.Branch            -> prettify()
-    is Control.Call              -> stringifyStatement("CALL", expression)
-    is Control.Raise             -> stringifyStatement("RAISE", expression)
+    is Control.Branch -> prettify()
+    is Control.Invoke -> stringifyStatement("CALL", expression)
+    is Control.Raise  -> stringifyStatement("RAISE", expression)
     is Control.Return            -> stringifyStatement("RETURN", expression)
-    is Function                  -> prettify()
+    is Definition.Function       -> prettify()
+    is Definition.Type           -> prettify()
+    is Definition.Variable       -> prettify()
     is Operator.Add              -> stringifyBinary("+", lhs, rhs)
     is Operator.And              -> stringifyBinary("&&", lhs, rhs)
     is Operator.Catch            -> stringifyBinary(":", lhs, rhs)
@@ -49,12 +50,9 @@ private fun Node.prettify(): String = when (this)
     is Operator.Subtract         -> stringifyBinary("-", lhs, rhs)
     is Operator.ThreeWay         -> stringifyBinary("<=>", lhs, rhs)
     is Operator.Xor              -> stringifyBinary("^^", lhs, rhs)
-    is Segment                   -> prettify()
-    is Type                      -> prettify()
     is Value.Bool                -> stringifyLiteral(value, null)
     is Value.Real                -> stringifyLiteral(value, type)
     is Value.Text                -> stringifyLiteral(value, type)
-    is Variable                  -> prettify()
     is When                      -> prettify()
 }
 
@@ -105,8 +103,6 @@ private fun stringifyLiteral(value: Any, type: Name?): String =
 private fun stringifyStatement(statement: String, expression: Expression?): String =
     "$statement ${expression.format { it.prettify() }}"
 
-// ...
-
 /**
  * Joins [this] collection of nodes into a single string. If [newlines] are assigned, all elements in [this] collection
  * are placed on a new line. If [indented] is also assigned, all entries are indented one additional level.
@@ -114,40 +110,44 @@ private fun stringifyStatement(statement: String, expression: Expression?): Stri
 private fun <T> Iterable<T>.prettify(newlines: Boolean, indented: Boolean, transformation: (T) -> String): String =
     joinToString(if (newlines) "\n" else ", ") { transformation(it) }.let { if (indented) it.indent() else it }
 
-private fun Segment.prettify(): String = """
-    MODULE ${module ?: "<none>"}
-    IMPORT ${imports.joinToString().ifBlank { "<none>" }}
-""".trimIndent() + "\n" + statements.prettify(newlines = true, indented = false) { it.prettify() }
+// ...
 
-private fun Parameter.prettify(): String =
-    "${name.format { "$it = " }}${expression.prettify()}"
+private fun Access.Function.prettify(): String =
+    "$name(${arguments.joinToString { it.prettify() }})"
 
-private fun Scope.prettify(): String =
-    "{\n" + statements.prettify(newlines = true, indented = true) { it.prettify() } + "\n}"
+private fun Access.Subscript.prettify(): String =
+    "$name[${arguments.joinToString { it.prettify() }}]"
 
-private fun Type.prettify(): String =
+private fun Control.Branch.prettify(): String =
+    "IF ${predicate.prettify()}\n${success.prettify()}${failure.format { "\nELSE\n${it.prettify()}" }}"
+
+private fun Definition.Function.prettify(): String =
+    "$visibility FUN $name(" + parameters.joinToString { it.prettify() } + ")\n{\n" + statements.prettify(newlines = true, indented = true) { it.prettify() } + "\n}"
+
+private fun Definition.Type.prettify(): String =
     "$visibility TYPE $name\n{\n" + properties.prettify(newlines = true, indented = true) { it.prettify() } + "\n}"
 
-private fun Type.Property.prettify(): String =
-    "$visibility $mutability $name${type.format { ": $it" }}${value.format { " = ${it.prettify()}" }}"
-
-private fun Variable.prettify(): String =
+private fun Definition.Variable.prettify(): String =
     "$visibility $mutability $name = ${value.prettify()}"
 
 private fun When.prettify(): String =
     "WHEN ${expression.prettify()}\n" + branches.joinToString("\n") { "${it.first.prettify()} -> ${it.second.prettify()}".indent() } + default.format { "\nELSE -> ${it.prettify()}".indent() }
 
-private fun Access.Function.prettify(): String =
-    "$name(${parameters.joinToString { it.prettify() }})"
+// ...
 
-private fun Access.Subscript.prettify(): String =
-    "$name[${parameters.joinToString { it.prettify() }}]"
+private fun Argument.prettify(): String =
+    "${name.format { "$it = " }}${expression.prettify()}"
 
-private fun Control.Branch.prettify(): String =
-    "IF ${predicate.prettify()}\n${success.prettify()}${failure.format { "\nELSE\n${it.prettify()}" }}"
+private fun Property.prettify(): String =
+    "$visibility $mutability $name${type.format { ": $it" }}${value.format { " = ${it.prettify()}" }}"
 
-private fun Function.prettify(): String =
-    "$visibility FUN $name(" + parameters.joinToString { it.prettify() } + ")\n" + scope.prettify()
-
-private fun Function.Parameter.prettify(): String =
+private fun Parameter.prettify(): String =
     "$mutability " + type.format { "$it " } + "$name${value.format { " = $it" }}"
+
+private fun Scope.prettify(): String =
+    "{\n" + statements.prettify(newlines = true, indented = true) { it.prettify() } + "\n}"
+
+private fun Segment.prettify(): String = """
+    MODULE ${module ?: "<none>"}
+    IMPORT ${imports.joinToString().ifBlank { "<none>" }}
+""".trimIndent() + "\n" + definitions.prettify(newlines = true, indented = false) { it.prettify() }
